feat: refactor language index creation into a dedicated script and update initialization process
