+++ backend/Dockerfile
@@ -5,7 +5,6 @@ COPY go.mod go.sum ./
 RUN go mod download
 
 COPY . ./
-RUN go run scripts/create_language_index.go
 RUN CGO_ENABLED=0 GOOS=linux go build -o server ./cmd/server
 
 FROM alpine:latest
