+++ backend/internal/scripts/runner.go
@@ -1,0 +1,13 @@
+package scripts
+
+import (
+	"log"
+)
+
+func RunAll() {
+	log.Println("Running initialization scripts...")
+	
+	CreateLanguageIndex()
+	
+	log.Println("All initialization scripts completed")
+}
